requirements.txt
flask
requests
beautifulsoup4
pandas
lxml
python-dotenv
